Issue 42: Add a "list" url for blog posts Links the REST URLs in the index page
