+++ config/routes.yaml
@@ -1,3 +1,3 @@
-index:
-    path: /
-    defaults: { _controller: 'App\Controller\IndexController' }
+#index:
+#    path: /
+#    controller: App\Controller\DefaultController::index
